-Added massive revamp to project. Made the folder structure more manageable.
