+++ Inventory-System/App.xaml
@@ -2,7 +2,7 @@
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:local="clr-namespace:Inventory_System"
-             StartupUri="UI\Main\MainWindow.xaml">
+             StartupUri="Common\MainWindow.xaml">
     <Application.Resources>
         <Style TargetType="TextBlock" x:Key="Header">
             <Setter Property="FontSize" Value="24"/>
