--- Inventory-System/DataAccess/Repositories/PC/PcRepository.cs
@@ -1,14 +1,0 @@
-﻿using Inventory_System.Data.Model;
-using System;
-using System.Collections.Generic;
-using System.Data.SqlClient;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Inventory_System.Data.Repositories.PC
-{
-    public class PcRepository
-    {
-    }
-}
