+++ Inventory-System/Features/SIMCard/Views/List/SIMCardView.xaml.cs
