+++ Inventory-System/Features/SIMType/Views/Create/MakeSIMCardTypeView.xaml.cs
