+++ Inventory-System/Features/SIMType/Views/List/SIMCardTypeListView.xaml
