+++ Inventory-System/Features/SIMType/Views/List/SIMCardTypeListViewModel.cs
@@ -1,0 +1,14 @@
+﻿using Inventory_System.Common;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Inventory_System.UI.Views.SIM.SIMCardType.View
+{
+    public class SIMCardTypeListViewModel : ViewModelBase
+    {
+  
+    }
+}
