+++ Inventory-System/Services/Configuration/SysConfig.cs
