--- Inventory-System/UI/Main/MainWindow.xaml
@@ -1,15 +1,0 @@
-﻿<Window x:Class="Inventory_System.MainWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:views="clr-namespace:Inventory_System.UI.Views.SIM.SIMCard.List"
-        mc:Ignorable="d"
-        Title="MainWindow" Height="450" Width="800">
-
-    <Grid>
-        <!--<views:MakeSIMCardTypeView Loaded="MakeSIMCardTypeView_Loaded" />-->
-        <!--<views:SIMCardTypeListView Margin="10,49,-10,-49"/>-->
-        <views:SIMCardView/>
-    </Grid>
-</Window>
