--- Inventory-System/UI/Views/SIM/SIMCard/List/SIMCardView.xaml
@@ -1,59 +1,0 @@
-﻿<UserControl x:Class="Inventory_System.UI.Views.SIM.SIMCard.List.SIMCardView"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:Inventory_System.UI.Views.SIM.SIMCard.List"
-             mc:Ignorable="d" 
-             d:DesignHeight="450" d:DesignWidth="800">
-
-    <UserControl.DataContext>
-        <local:SIMCardViewModel/>
-    </UserControl.DataContext>
-    
-    <Grid>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="Auto"/>
-        </Grid.RowDefinitions>
-
-        <Grid Grid.Row="0">
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition/>
-                <ColumnDefinition/>
-            </Grid.ColumnDefinitions>
-
-            <TextBlock 
-                Text="SIM Cards"
-                Grid.ColumnSpan="2"
-                Margin="0,5,0,0"
-                Style="{StaticResource Header}"/>
-
-            <Button 
-                Content="Create"
-                Style="{StaticResource CreateButton}"
-                Grid.Column="1"
-                HorizontalAlignment="Right"/>
-        </Grid>
-        <StackPanel Grid.Row="1" Height="Auto" HorizontalAlignment="Stretch">
-            <DataGrid
-                Margin="20"
-                Height="250px"
-                ColumnWidth="*"
-                HeadersVisibility="Column"
-                AutoGenerateColumns="False"
-                CanUserAddRows="False"
-                ItemsSource="{Binding SIMCards}">
-                <DataGrid.Columns>
-                    <DataGridTextColumn Header="ID" Binding="{Binding id}"/>
-                    <DataGridTextColumn Header="Number" Binding="{Binding number}"/>
-                    <DataGridTextColumn Header="PIN" Binding="{Binding PIN}"/>
-                    <DataGridTextColumn Header="PUK" Binding="{Binding PUK}"/>
-                    <DataGridTextColumn Header="ICCID" Binding="{Binding ICCID}"/>
-                    <DataGridTextColumn Header="SIM Type" Binding="{Binding SIMTYPE.simType}" />
-                </DataGrid.Columns>
-            </DataGrid>
-        </StackPanel>
-       
-    </Grid>
-</UserControl>
