--- Inventory-System/UI/Views/SIM/SIMCardType/View/SIMCardTypeListViewModel.cs
@@ -1,15 +1,0 @@
-﻿using Inventory_System.UI.Common;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-using Inventory_System.Data.Model.Device.SIM_card;
-namespace Inventory_System.UI.Views.SIM.SIMCardType.View
-{
-    public class SIMCardTypeListViewModel : ViewModelBase
-    {
-  
-    }
-}
